+++ src/funtracks/data_model/__init__.py
@@ -1,2 +1,1 @@
 from .tracks import Tracks  # noqa
-from .solution_tracks import SolutionTracks  # noqa
