+++ src/funtracks/import_export/_tracks_builder.py
@@ -1,6 +1,6 @@
 """Builder pattern for importing tracks from various formats.
 
-This module provides a unified interface for constructing SolutionTracks objects
+This module provides a unified interface for constructing Tracks objects
 from different data sources (GEFF, CSV, etc.) while sharing common validation
 and construction logic.
 """
@@ -15,7 +15,7 @@ import numpy as np
 import tracksdata as td
 from geff._typing import InMemoryGeff
 
-from funtracks.data_model.solution_tracks import SolutionTracks
+from funtracks.data_model.tracks import Tracks
 from funtracks.features import Feature
 from funtracks.import_export._import_segmentation import (
     load_segmentation,
@@ -610,13 +610,13 @@ class TracksBuilder(ABC):
 
         return new_segmentation, scale, graph
 
-    # Structural keys that are handled by graph construction / SolutionTracks.__init__
+    # Structural keys that are handled by graph construction / Tracks.__init__
     # and should not be registered as features by enable_features().
     STRUCTURAL_KEYS = frozenset({"time", "id", "parent_id", "seg_id"})
 
     def enable_features(
         self,
-        tracks: SolutionTracks,
+        tracks: Tracks,
         name_map: dict[str, str | list[str]],
         feature_type: Literal["node", "edge"] = "node",
     ) -> None:
@@ -629,7 +629,7 @@ class TracksBuilder(ABC):
         - Otherwise, if data was loaded for it, register it as a static feature.
 
         Args:
-            tracks: SolutionTracks object to add features to
+            tracks: Tracks object to add features to
             name_map: Mapping from standard funtracks keys to source property
                 names (same format as node_name_map / edge_name_map).
             feature_type: Type of features ("node" or "edge")
@@ -676,7 +676,7 @@ class TracksBuilder(ABC):
         scale: list[float] | None = None,
         node_name_map: dict[str, str | list[str]] | None = None,
         database: str | None = None,
-    ) -> SolutionTracks:
+    ) -> Tracks:
         """Orchestrate the full construction process.
 
         Args:
@@ -688,7 +688,7 @@ class TracksBuilder(ABC):
                 If None (default), an in-memory/temp graph is used.
 
         Returns:
-            Fully constructed SolutionTracks object
+            Fully constructed Tracks object
 
         Raises:
             ValueError: If self.node_name_map is not set or validation fails
@@ -760,22 +760,36 @@ class TracksBuilder(ABC):
         if segmentation_array is not None:
             graph = add_masks_and_bboxes_to_graph(graph, segmentation_array)
 
-        # 7. Create SolutionTracks
+        # 7. Create Tracks
         # construct_graph() always stores time as "t" (tracksdata convention),
         # regardless of TIME_ATTR, so we pass "t" here explicitly.
         # If a FeatureDict was loaded (e.g., from GEFF metadata), use it directly
         if hasattr(self, "features") and self.features is not None:
-            tracks = SolutionTracks(
+            tracks = Tracks(
                 graph=graph,
                 ndim=self.ndim,
                 scale=scale,
                 features=self.features,
             )
         else:
-            tracks = SolutionTracks(
+            # The builder always produces a solution, so declare tracklet/lineage
+            # intent to register a TrackAnnotator. Use the attr name present on the
+            # constructed graph; fall back to the default keys (computed from scratch)
+            # when the source carried no track ids.
+            node_keys = graph.node_attr_keys()
+            tracklet_attr = next(
+                (k for k in ("tracklet_id", "track_id") if k in node_keys),
+                "tracklet_id",
+            )
+            lineage_attr = next(
+                (k for k in ("lineage_id",) if k in node_keys), "lineage_id"
+            )
+            tracks = Tracks(
                 graph=graph,
                 pos_attr="pos",
                 time_attr="t",
+                tracklet_attr=tracklet_attr,
+                lineage_attr=lineage_attr,
                 ndim=self.ndim,
                 scale=scale,
             )
