+++ src/funtracks/import_export/_v1_format.py
@@ -16,7 +16,7 @@ from funtracks.utils.tracksdata_utils import (
 )
 
 if TYPE_CHECKING:
-    from ..data_model import SolutionTracks, Tracks
+    from ..data_model import Tracks
 
 GRAPH_FILE = "graph.json"
 SEG_FILE = "seg.npy"
@@ -25,7 +25,7 @@ ATTRS_FILE = "attrs.json"
 
 def load_v1_tracks(
     directory: Path, seg_required: bool = False, solution: bool = False
-) -> Tracks | SolutionTracks:
+) -> Tracks:
     """Load a Tracks object from the given directory. Looks for files
     in the format generated by Tracks.save.
 
@@ -35,7 +35,7 @@ def load_v1_tracks(
         directory (Path): The directory containing tracks to load
         seg_required (bool, optional): If true, raises a FileNotFoundError if the
             segmentation file is not present in the directory. Defaults to False.
-        solution (bool, optional): If true, returns a SolutionTracks object, otherwise
+        solution (bool, optional): If true, returns a Tracks object, otherwise
             returns a normal Tracks object. Defaults to False.
 
     Returns:
@@ -114,17 +114,20 @@ def load_v1_tracks(
     # is breaking, and manually setting the attrs to None if features is present will
     # break if the attrs are changed/removed in the future. Can remove in v2.0.
     # Import at runtime to avoid circular dependency
-    from ..data_model import SolutionTracks, Tracks
+    from ..data_model import Tracks
+
+    # A solution save must come back with track ids. When the stored attrs don't carry
+    # a FeatureDict (older saves), declare tracklet/lineage intent explicitly so a
+    # TrackAnnotator is registered.
+    if solution and "features" not in attrs:
+        attrs.setdefault("tracklet_attr", "tracklet_id")
+        attrs.setdefault("lineage_attr", "lineage_id")
 
     with warnings.catch_warnings():
         warnings.filterwarnings(
             "ignore", message="Provided both FeatureDict and pos_attr or time_attr"
         )
-        tracks: Tracks
-        if solution:
-            tracks = SolutionTracks(graph_td, **attrs)
-        else:
-            tracks = Tracks(graph_td, **attrs)
+        tracks = Tracks(graph_td, **attrs)
     return tracks
