+++ src/funtracks/user_actions/user_delete_edge.py
@@ -9,19 +9,19 @@ from ..actions.add_delete_edge import DeleteEdge
 from ..actions.update_track_id import UpdateTrackIDs
 
 if TYPE_CHECKING:
-    from funtracks.data_model import SolutionTracks
+    from funtracks.data_model import Tracks
 
 
 class UserDeleteEdge(ActionGroup):
     def __init__(
         self,
-        tracks: SolutionTracks,
+        tracks: Tracks,
         edge: tuple[int, int],
         _top_level: bool = True,
     ):
         """
         Args:
-            tracks (SolutionTracks): The tracks to delete the edge from.
+            tracks (Tracks): The tracks to delete the edge from.
             edge (tuple[int, int]): The edge to delete.
             _top_level (bool): If True, add this action to the history and emit
                 refresh. Set to False when used as a sub-action inside a compound
@@ -31,7 +31,7 @@ class UserDeleteEdge(ActionGroup):
             InvalidActionError: If the edge does not exist in the graph.
         """
         super().__init__(tracks, actions=[])
-        self.tracks: SolutionTracks  # Narrow type from base class
+        self.tracks: Tracks  # Narrow type from base class
         if not self.tracks.graph_solution.has_edge(*edge):
             raise InvalidActionError(f"Edge {edge} not in solution, can't remove")
