+++ tests/annotators/test_edge_annotator.py
@@ -4,7 +4,7 @@ from tracksdata.nodes import Mask
 
 from funtracks.actions import UpdateNodeSeg, UpdateTrackIDs
 from funtracks.annotators import EdgeAnnotator
-from funtracks.data_model import SolutionTracks, Tracks
+from funtracks.data_model import Tracks
 
 track_attrs = {"time_attr": "t", "tracklet_attr": "track_id"}
 
@@ -134,7 +134,7 @@ class TestEdgeAnnotator:
     def test_ignores_irrelevant_actions(self, get_graph, ndim):
         """Test that EdgeAnnotator ignores actions that don't affect edges."""
         graph = get_graph(ndim, is_solution=True, with_seg=True)
-        tracks = SolutionTracks(
+        tracks = Tracks(
             graph,
             ndim=ndim,
             **track_attrs,
